PDF Thumbnail added

+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"bufio"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -25,14 +26,15 @@ var app *fiber.App
 
 func main() {
 	println(version)
-	// println(time.Now().Unix())
-	// if time.Now().Unix() > 1713262858 {
-	// 	println("Please Update The Binary From Keyur Shah")
-	// 	println("Press Any Key To Close")
-	// 	input := bufio.NewScanner(os.Stdin)
-	// 	input.Scan()
-	// 	return
-	// }
+
+	if time.Now().Unix() > 1719562963 {
+		println("Please Update The Binary From Keyur Shah")
+		println("Press Any Key To Close")
+		input := bufio.NewScanner(os.Stdin)
+		input.Scan()
+		return
+	}
+
 	env.CurrentDirectory = FindAndReturnCurrentDir()
 	go func() {
 		os.RemoveAll("./tmp")
